+++ doc/changelog.md
@@ -4,13 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-# 1.1.0 - IN DEV
+# 1.1.0 - 2017-03-14
 ### Added
 * Database class can now import from compressed archive.
 * File class can now perform deployment or withdrawal.
 * New Composer custom installer for Wordpress and Drupal CMS, you can now use composer type `rocket-cms` to migrate a package to cms folder. ( default : `web/edition/` ).
+* Added Composer checking version.
 ### Changed
-* Application environnment is now set to production by default
+* Composer System Installer is now deprecated and replaced by Composer Installer which is now placed on boilerplate.
+* Application environment is now set to production by default
 * Clean function has been removed from composer install
 * Complete code review for standard coding syntax.